requirements.txt
nextcord
pymongo
dnspython
python-dotenv
aiohttp